refactor derive mod
