+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "artimonist-cli"
-version = "1.0.0"
+version = "1.1.0"
 description = "A tool for generating mnemonics and wallets."
 authors = ["Artimonist <artimonist@gmail.com>"]
 homepage = "https://artimonist.org/"
@@ -25,8 +25,9 @@ artimonist = { version = "0.11.5", features = ["multilingual"] }
 clap = { version = "^4.5.28", features = ["derive"] }
 inquire = { version = "^0.7.5", default-features = false, features = ["crossterm"] }
 comfy-table = { version = "^7.1.4", default-features = false }
-thiserror = "2.0.11"
 bip38 = "1.1.1"
+thiserror = "2.0.11"
+anyhow = "1.0.98"
 
 [profile.release]
 codegen-units = 1
